move average_models to onmt.bin, create entry point
